requirements.txt
sqlalchemy==2.0.8
psycopg2-binary==2.9.5
alembic==1.10.2
pika==1.3.1
ansible-runner==2.3.2
python-dotenv==1.0. 0